Remove redundant casts flagged by phpstan-strict-rules
Strip 184 (int)/(float)/(string)/(bool) cast operators that operate on a
value already of the target type, across the clustering, selection, feed,
metadata, thumbnail and monitoring layers. Behaviour is unchanged: PHPStan
proved each cast a no-op. Includes the two knock-on simplifications the
removals exposed (a mirror self-assignment and now-redundant parentheses).

+++ src/Clusterer/DaySummaryStage/AwayFlagStage.php
@@ -220,8 +220,8 @@ final readonly class AwayFlagStage implements DaySummaryStageInterface
         $requiredCover = (int) ($windowSeconds * 0.75);
 
         foreach ($staypoints as $staypoint) {
-            $stayStart = (int) ($staypoint['start'] ?? 0);
-            $stayEnd   = (int) ($staypoint['end'] ?? 0);
+            $stayStart = ($staypoint['start'] ?? 0);
+            $stayEnd   = ($staypoint['end'] ?? 0);
             if ($stayStart === 0) {
                 continue;
             }
@@ -258,8 +258,8 @@ final readonly class AwayFlagStage implements DaySummaryStageInterface
         $endTs   = $windowEnd->getTimestamp();
 
         foreach ($staypoints as $staypoint) {
-            $stayStart = (int) ($staypoint['start'] ?? 0);
-            $stayEnd   = (int) ($staypoint['end'] ?? 0);
+            $stayStart = ($staypoint['start'] ?? 0);
+            $stayEnd   = ($staypoint['end'] ?? 0);
             if ($stayEnd < $startTs) {
                 continue;
             }
